Refactor build script and update dependencies
This commit includes the following changes:

- Removed unused dependencies from Cargo.toml, including `reqwest`, `flate2`, `tar`, `zip`, `anyhow`, `dirs`, and `bump2version`.
- Simplified the build.rs script by removing the code that downloaded and extracted external dependencies. This functionality is no longer needed as the project now relies on the Rust standard library and the `pyo3` crate for its dependencies.
- Updated the `pyo3` crate to version 0.22.2 and enabled the `extension-module` and `generate-import-lib` features.
- Removed the commented-out configuration for the `maturin` package metadata, as it is no longer relevant.

The main reason for these changes is to simplify the project's build process and reduce the number of external dependencies. By removing the code that downloaded and extracted external dependencies, the build script becomes more straightforward and easier to maintain. Additionally, the updated `pyo3` crate version provides improved functionality and stability for the project's Python bindings.